+++ rest/web_landing.go
@@ -1,0 +1,20 @@
+package rest
+
+import (
+	"net/http"
+	"os"
+
+	"github.com/gin-gonic/gin"
+)
+
+func pingPong(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{
+		"message": "pong",
+	})
+}
+
+func version(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{
+		"message": os.Getenv("VERSION"),
+	})
+}